+++ R/sample_pseudoabs.R
@@ -1,26 +1,26 @@
-#' Sample pseudo-absence (or background) points for SDM analysis
+#' Sample pseudo-absence points for SDM analysis
 #'
-#' This function samples pseudo-absence (or background, the naming is a matter
-#' of semantics) points from a raster given a set of presences.
+#' This function samples pseudo-absence points from a raster given a set of presences.
 #' The locations returned as the center points of the sampled cells, which can
-#' not overlap with the presences. The following methods are implemented:
-#' * 'random': pseudo-absences/background randomly sampled from the region covered by the
+#' not overlap with the presences (in contrast to background points, see 
+#' [sample_background]). The following methods are implemented:
+#' * 'random': pseudo-absences randomly sampled from the region covered by the
 #' raster (i.e. not NAs).
-#' * 'dist_min': pseudo-absences/background randomly sampled from the region excluding a buffer
+#' * 'dist_min': pseudo-absences randomly sampled from the region excluding a buffer
 #' of 'dist_min' from presences (distances in 'm' for lonlat rasters, and in map
 #' units for projected rasters).
-#' * 'dist_max': pseudo-absences/background randomly sampled from the unioned buffers
+#' * 'dist_max': pseudo-absences randomly sampled from the unioned buffers
 #' of 'dist_max' from presences (distances in 'm' for lonlat rasters, and in map
 #' units for projected rasters). Using the union of buffers means that areas that
 #' are in multiple buffers are not oversampled. This is also referred to as "thickening".
-#' * 'dist_disc': pseudo-absences/background randomly sampled from the unioned discs around presences
+#' * 'dist_disc': pseudo-absences randomly sampled from the unioned discs around presences
 #' with the two values of 'dist_disc' defining the minimum and maximum distance from
 #' presences.
 #' @param data An [`sf::sf`] data frame, or a data frame with coordinate variables.
 #' These can be defined in `coords`, unless they have standard names
 #' (see details below).
 #' @param raster the [terra::SpatRaster] from which cells will be sampled
-#' @param n number of pseudoabsence/background points to sample
+#' @param n number of pseudoabsence points to sample
 #' @param coords a vector of length two giving the names of the "x" and "y"
 #' coordinates, as found in `data`. If left to NULL, the function will
 #' try to guess the columns based on standard names `c("x", "y")`, `c("X","Y")`,
@@ -29,7 +29,7 @@
 #' 'dist_disc'. Threshold distances are set as additional elements of a vector,
 #' e.g c('dist_min',70000) or c('dist_disc',50000,200000).
 #' @param class_label the label given to the sampled points. Defaults to `pseudoabs`
-#' @param return_pres return presences together with pseudoabsences/background
+#' @param return_pres return presences together with pseudoabsences
 #'  in a single tibble
 #' @returns An object of class [tibble::tibble]. If presences are returned, the
 #' presence level is set as the reference (to match the expectations in the