+++ R/tidysdm-package.R
@@ -1,0 +1,10 @@
+#' @keywords internal
+"_PACKAGE"
+
+## usethis namespace: start
+## usethis namespace: end
+#' @import methods
+#' @import tidymodels
+#' @import spatialsample
+#' @importFrom rlang .data
+NULL