+++ data-raw/make_data/create_lacerta_data.R
@@ -18,6 +18,19 @@ distrib <- readr::read_delim(file.path(tempdir(), "0068808-230530130749713.csv")
 lacerta <- distrib[, c("gbifID", "decimalLatitude", "decimalLongitude")]
 names(lacerta) <- c("ID", "latitude", "longitude")
 
+# quality checks and filter and exclude outliers 
+# (it may take some time, depending on the size of the dataset):
+library(CoordinateCleaner)
+
+lacerta_df <- data.frame(lacerta)
+
+flags <- clean_coordinates(x = lacerta_df,
+                           lon = "longitude",
+                           lat = "latitude",
+                           species = "ID")
+
+lacerta <- lacerta_df[flags$.summary,]
+
 usethis::use_data(lacerta, overwrite = TRUE)
 # saveRDS(lacerta, file="./inst/extdata/lacerta_coords.RDS")
 
@@ -68,10 +81,10 @@ vars_uncor <- c("bio15", "bio05", "bio13", "bio06")
 climate_future <- pastclim::region_slice(
   time_ce = 2090,
   bio_variables = vars_uncor,
-  data = "WorldClim_2.1_HadGEM3-GC31-LL_ssp245_10m",
+  dataset = "WorldClim_2.1_HadGEM3-GC31-LL_ssp245_10m",
   crop = iberia_poly
 )
-writeCDF(climate_present, "./inst/extdata/lacerta_climate_future_10m.nc",
+writeCDF(climate_future, "./inst/extdata/lacerta_climate_future_10m.nc",
   compression = 9, split = TRUE, overwrite = TRUE
 )
 # fix time axis (this is a workaround if we open the file with sf)
@@ -80,3 +93,18 @@ ncdf4::ncatt_put(nc_in, varid = "time", attname = "axis", attval = "T")
 ncdf4::nc_close(nc_in)
 
 #####
+# from the vignette, we save
+# lacerta_thin (present, absences plus all climatic variables)
+# lacerta ensemble
+usethis::use_data(lacerta_ensemble, overwrite=TRUE)
+
+################################
+occ_download_get(key = "0121761-240321170329656", path = tempdir())
+# read file
+library(readr)
+backg_distrib <- readr::read_delim(file.path(tempdir(), "0121761-240321170329656.zip"))
+
+# keep the necessary columns
+lacertidae_background <- backg_distrib %>% select(gbifID, decimalLatitude, decimalLongitude) %>%
+  rename(ID = gbifID, latitude = decimalLatitude, longitude = decimalLongitude)
+usethis::use_data(lacertidae_background, overwrite=TRUE)